Updating projects build & upload separated jars
