--- Core/src/main/java/com/microsoft/applicationinsights/logging/common/TelemetryClientProxy.java
@@ -1,24 +1,0 @@
-package com.microsoft.applicationinsights.logging.common;
-
-import com.microsoft.applicationinsights.channel.TelemetryClient;
-
-public interface TelemetryClientProxy {
-
-    /**
-     * Sends the given event to AI.
-     *
-     * @param event Event to send.
-     */
-    void sendEvent(ApplicationInsightsEvent event);
-
-    /**
-     * Gets the telemetry client.
-     */
-    TelemetryClient getTelemetryClient();
-
-    /**
-     * Gets a value indicating whether the proxy has been initialized.
-     * @return True if initialized, false otherwise.
-     */
-    boolean isInitialized();
-}
