--- Core/src/test/resources/log4j.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-    <appender name="test" class="com.microsoft.applicationinsights.logging.log4j.v1_2.ApplicationInsightsAppender">
-        <param name="instrumentationKey" value="Test_Instrumentation_Key" />
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%-5p %c{1} - %m%n"/>
-        </layout>
-    </appender>
-    <root>
-        <priority value ="trace" />
-        <appender-ref ref="test" />
-    </root>
-</log4j:configuration>
