+++ Logging/src/test/resources/applicationinsights.properties
@@ -1,0 +1,1 @@
+ai.instrumentationKey: THIS_PROPERTY_IS_MANDATORY
