+++ Logging/src/test/resources/logback.xml
@@ -1,0 +1,9 @@
+<configuration>
+    <appender name="test" class="com.microsoft.applicationinsights.logback.ApplicationInsightsAppender">
+        <instrumentationKey>Test_Instrumentation_Key</instrumentationKey>
+    </appender>
+
+    <root level="trace">
+        <appender-ref ref="test" />
+    </root>
+</configuration>
