+++ settings.gradle
@@ -1,4 +1,4 @@
 rootProject.name = 'AppInsights Java'
 
-include 'Core'
+include 'Core', 'Logging'
